chore(a2-2155): move use-existing pages into BYS directory
