+++ packages/forms-web-app/src/controllers/before-you-start/use-existing-service-application-type.js
@@ -2,7 +2,7 @@ const { VIEW } = require('../../lib/views');
 const config = require('../../config');
 
 exports.getUseExistingServiceApplicationType = async (_, res) => {
-	res.render(VIEW.FULL_APPEAL.USE_EXISTING_SERVICE_APPLICATION_TYPE, {
+	res.render(VIEW.BEFORE_YOU_START.USE_EXISTING_SERVICE_APPLICATION_TYPE, {
 		bannerHtmlOverride: config.betaBannerText,
 		acpLink: 'https://acp.planninginspectorate.gov.uk/'
 	});
